main/requirements.txt
redis==5.2.1
fastapi==0.95.0
uvicorn==0.21.1
python-socketio==5.13.0

